% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{get_artifacts}
\alias{get_artifacts}
\title{Get manual}
\usage{
get_artifacts(view, branch)
}
\arguments{
\item{view}{data.frame package information}

\item{branch}{character "devel" or "release"}
}
\description{
Get manual
}
