% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{get_citation}
\alias{get_citation}
\title{Get citation}
\usage{
get_citation(pkg_url, save_path = NULL, ext = "html")
}
\arguments{
\item{pkg_url}{character R Universe URL for the package}

\item{save_path}{character (default: NULL) save path}

\item{ext}{character (default: html)}
}
\description{
Get citation
}
\examples{
pkg_url <- .runiverse_url("BiocCheck", "devel")
get_citation(pkg_url)

}
