% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{get_manual}
\alias{get_manual}
\title{Get manual}
\usage{
get_manual(pkg, pkg_url, save_path = NULL, ext = c("pdf", "html"))
}
\arguments{
\item{pkg}{character package name}

\item{pkg_url}{character R Universe URL for the package}

\item{save_path}{character (default: NULL) save path}

\item{ext}{character (default: c("pdf", "html"))}
}
\description{
Get manual
}
\details{
Packages should be placed along the following paths where pkg is
the package name:
\itemize{
\item{"manuals/pkg/refman/pkg.html"}
\item{"manuals/pkg/man/pkg.pdf"}
\item{"web/packages/pkg/refman/pkg.html"}
\item{"web/packages/pkg/pkg.pdf"}
}
}
\examples{
pkg_url <- .runiverse_url("BiocCheck", "devel")
get_manual(pkg, pkg_url, ext = "pdf")

}
