% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{get_news}
\alias{get_news}
\title{Get NEWS}
\usage{
get_news(pkg_url, save_path = NULL)
}
\arguments{
\item{pkg_url}{character R Universe URL for the package}

\item{save_path}{character (default: NULL) save path}
}
\description{
Get NEWS
}
\examples{
pkg_url <- .runiverse_url("BiocCheck", "devel")
get_news(pkg_url)

}
