% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{get_vignettes}
\alias{get_vignettes}
\title{Get vignettes}
\usage{
get_vignettes(pkg_url, vignettes, save_path = NULL)
}
\arguments{
\item{pkg_url}{character R Universe URL for the package}

\item{vignettes}{list paths to vignettes}

\item{save_path}{character (default: NULL) save path}
}
\description{
Get vignettes
}
\examples{
pkg_url <- .runiverse_url("BiocCheck", "devel")
df <- get_ruData("BiocCheck", "devel")
view <- prepareView(df) 
get_vignettes(pkg_url, view$vignettes)

}
