% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{passed_criteria}
\alias{passed_criteria}
\title{Passed packages VIEWS update}
\usage{
passed_criteria(views, universe, r_version, os = c("linux", "win", "mac"))
}
\arguments{
\item{views}{data.frame representing VIEWS}

\item{universe}{character corresponding R Universe}

\item{r_version}{character R version}

\item{os}{(default: linux, win, mac) OSes required to pass}
}
\value{
data.frame, one row per package source or binary
}
\description{
Passed packages VIEWS update
}
\examples{
bu <- biocUniTools::uni_for_bioc("release")
views <- read_package_views(c("BiocCheck", "bedbaser"), "release",
    "software")
passed <- passed_criteria(views, bu$universe, bu$r_version)

}
