% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{read_package_views}
\alias{read_package_views}
\title{Read all package views}
\usage{
read_package_views(branch, package_type)
}
\arguments{
\item{branch}{character "release" or "devel"}

\item{package_type}{character a value in in .PACKAGE_TYPES}
}
\value{
data.frame, all package view data, with list-columns for
fields such as Suggests/Depends/Imports/biocViews that have
multiple values in at least one package
}
\description{
Read all package views
}
\examples{
views <- read_package_views("devel", "software")

}
