% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{simplify_columns}
\alias{simplify_columns}
\title{Simplify list-columns back to atomic vectors where every row's
value is actually a length-1 atomic value; leave real list-columns
(multi-valued fields) untouched}
\usage{
simplify_columns(df)
}
\arguments{
\item{df}{data.frame with list-columns from bind_rows(view_to_row(...))}
}
\value{
data.frame
}
\description{
Simplify list-columns back to atomic vectors where every row's
value is actually a length-1 atomic value; leave real list-columns
(multi-valued fields) untouched
}
