% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{update_package_type_views}
\alias{update_package_type_views}
\title{Update the views for passed packages of a package_type}
\usage{
update_package_type_views(
  manifest_url = .MANIFEST_URL,
  os = .OS,
  branch = .BRANCHES,
  package_type = .PACKAGE_TYPES
)
}
\arguments{
\item{manifest_url}{character (default: .MANIFEST_URL)}

\item{os}{character or vector of "linux", "mac", "win", etc}

\item{branch}{character or vector in .BRANCHES}

\item{package_type}{character or vector in .PACKAGE_TYPES}
}
\value{
vector of characters paths to the package_type VIEWS
}
\description{
Update the views for passed packages of a package_type
}
\examples{
update_package_type_views()

}
