% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views_new.R
\name{update_package_views}
\alias{update_package_views}
\title{Update the views for passed packages of a package_type}
\usage{
update_package_views(
  passed,
  branch,
  bioc_version,
  package_type,
  verbose = FALSE
)
}
\arguments{
\item{passed}{data.frame of passed package information}

\item{branch}{character "release" or "devel"}

\item{bioc_version}{character Bioconductor version}

\item{package_type}{character}

\item{verbose}{logical (default: FALSE)}
}
\value{
integer number of views updated
}
\description{
Update the views for passed packages of a package_type
}
