% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{write_package_views}
\alias{write_package_views}
\title{Update the views for passed packages of a package_type}
\usage{
write_package_views(
  packages_df,
  branch,
  bioc_version,
  package_type,
  verbose = FALSE
)
}
\arguments{
\item{packages_df}{data.frame of passed package information}

\item{branch}{character "release" or "devel"}

\item{bioc_version}{character Bioconductor version}

\item{package_type}{character}

\item{verbose}{logical (default: FALSE)}
}
\value{
integer number of views updated
}
\description{
Update the views for passed packages of a package_type
}
