% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views_new.R
\name{write_propagation_status}
\alias{write_propagation_status}
\title{Write propagation status for a package type}
\usage{
write_propagation_status(status, branch, package_type)
}
\arguments{
\item{status}{data.frame, output of \code{passed_criteria()}}

\item{branch}{character "release" or "devel"}

\item{package_type}{character}
}
\value{
character, the path written to
}
\description{
Writes out which \verb{(Package, OS, Arch)} combinations currently meet
propagation criteria -- i.e. which binaries should be made publicly
available -- as a SEPARATE artifact from the VIEWS files themselves.
This is meant to be consumed by another process to decide what to
actually publish/fetch (see the original project description: other
URLs get collected separately for that purpose).
}
