% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{changed_packages}
\alias{changed_packages}
\title{Identify new R Universe builds}
\usage{
changed_packages(views, universe_df)
}
\arguments{
\item{views}{data.frame views corresponding to a branch}

\item{universe_df}{data.frame universe data corresponding to branch
must have Package and RemoteSha columns}
}
\value{
character vector of package names that need updating
}
\description{
Compares each package's \code{git_last_commit} against R-Universe's \code{RemoteSha}
Packages with no view are treated as changed.
}
\examples{
views <- read_package_views(c("BiocCheck", "a4Reporting"),"devel",
    "software")
universe_df <- biocUniTools::get_raw_uni_df("bioc")
changed <- changed_packages(views, universe_df)

}
