% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{package_view}
\alias{package_view}
\title{Prepare package VIEW data}
\usage{
package_view(df, git_branch)
}
\arguments{
\item{df}{data.frame R Universe api data}

\item{git_branch}{character "devel" or "RELEASE_X_Y"}
}
\value{
a list of named fields for a package VIEW entry in the VIEWS
file
}
\description{
Prepare the VIEW of a package. Does not include
source, binary, extra doc, Rfiles, and reverse dependency fields.
}
\examples{
bu <- biocUniTools::uni_for_bioc("devel")
df <- biocUniTools::get_jobs(bu)
git_branch <- .get_branch(bu$bioc_branch, bu$bioc_version)
view <- package_view(df[1, ], git_branch)

}
