% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{passed_packages}
\alias{passed_packages}
\title{Filter for packages passing R-Universe R CMD check}
\usage{
passed_packages(packages, universe, os = c("linux", "win", "mac"))
}
\arguments{
\item{packages}{vector packages that were updated recently}

\item{universe}{character corresponding R Universe}

\item{os}{character (default: linux, win, mac) OSes required to pass}
}
\value{
data.frame Packages passing on an OS
}
\description{
Filter for packages passing R-Universe R CMD check
}
\examples{
bu <- biocUniTools::uni_for_bioc("release")
packages <- c("BiocCheck", "BiocFileCache")
passed <- passed_packages(packages, bu$universe)

}
