% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{read_view}
\alias{read_view}
\title{Read a single package view from JSON file}
\usage{
read_view(package, path)
}
\arguments{
\item{package}{character package name}

\item{path}{character}
}
\value{
list, the package view data
}
\description{
Read a single package view from JSON file
}
\examples{
view <- read_view("devel/software", "BiocCheck")

}
