% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{write_package_views}
\alias{write_package_views}
\title{Write package views}
\usage{
write_package_views(
  packages_df,
  branch,
  bioc_version,
  package_type,
  verbose = FALSE
)
}
\arguments{
\item{packages_df}{data.frame of passed package information}

\item{branch}{character "release" or "devel"}

\item{bioc_version}{character Bioconductor version}

\item{package_type}{character}

\item{verbose}{logical (default: FALSE)}
}
\value{
integer number of views updated
}
\description{
Write package views
}
\examples{
bu <- biocUniTools::uni_for_bioc("devel")
packages_df <- biocUniTools::get_raw_uni_df(bu$universe)
write_package_views(packages_df, bu$bioc_branch, bu$bioc_version, "software")

}
