% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{write_view}
\alias{write_view}
\title{Write view}
\usage{
write_view(df, save_path, ext = c("json", "dcf"))
}
\arguments{
\item{df}{data.frame package data}

\item{save_path}{character path}

\item{ext}{character json or dcf}
}
\description{
Write view
}
\examples{
bu <- biocUniTools::uni_for_bioc("devel")
df <- biocUniTools::get_jobs(bu)
view <- package_view(df[1, ], bu)
write_view(view, tempfile())

}
